Trait cordic::CordicNumber

pub trait CordicNumber: Copy + PartialOrd + AddAssign + SubAssign + Div<Output = Self> + Mul<Output = Self> + Neg<Output = Self> + Sub<Output = Self> + Add<Output = Self> + Shr<u8, Output = Self> + Shl<u8, Output = Self> {
    fn floor(self) -> Self;
    fn zero() -> Self;
    fn one() -> Self;
    fn frac_pi_2() -> Self;
    fn pi() -> Self;
    fn e() -> Self;
    fn from_u0f64(val: U0F64) -> Self;
    fn num_fract_bits() -> u8;
    fn num_bits() -> u8;

    fn half() -> Self { ... }
}

A number that can be used by the CORDIC-based algorithms.

This covers most fixed-point numbers, with some restriction on the maximal number of decimal bits in order to allow some constraints (like PI) to fit.

Required Methods

fn floor(self) -> Self

fn zero() -> Self

fn one() -> Self

fn frac_pi_2() -> Self

fn pi() -> Self

fn e() -> Self

fn from_u0f64(val: U0F64) -> Self

fn num_fract_bits() -> u8

fn num_bits() -> u8

Provided Methods

fn half() -> Self